+++ gr/gr_mod_4fsk_sdr.cpp
@@ -90,7 +90,7 @@ gr_mod_4fsk_sdr::gr_mod_4fsk_sdr(int sps, int samp_rate, int carrier_freq,
 
 }
 
-void gr_mod_4fsk_sdr::set_bb_gain(int value)
+void gr_mod_4fsk_sdr::set_bb_gain(float value)
 {
     _bb_gain->set_k(value);
 }
