+++ gr/gr_mod_am_sdr.cpp
@@ -83,7 +83,7 @@ void gr_mod_am_sdr::set_filter_width(int filter_width)
     _filter->set_taps(filter_taps);
 }
 
-void gr_mod_am_sdr::set_bb_gain(int value)
+void gr_mod_am_sdr::set_bb_gain(float value)
 {
     _bb_gain->set_k(value);
 }
