+++ gr/gr_mod_bpsk_sdr.cpp
@@ -83,7 +83,7 @@ gr_mod_bpsk_sdr::gr_mod_bpsk_sdr(int sps, int samp_rate, int carrier_freq,
 
 }
 
-void gr_mod_bpsk_sdr::set_bb_gain(int value)
+void gr_mod_bpsk_sdr::set_bb_gain(float value)
 {
     _bb_gain->set_k(value);
 }
