+++ gr/gr_mod_qpsk_sdr.cpp
@@ -100,7 +100,7 @@ gr_mod_qpsk_sdr::gr_mod_qpsk_sdr(int sps, int samp_rate, int carrier_freq,
 
 }
 
-void gr_mod_qpsk_sdr::set_bb_gain(int value)
+void gr_mod_qpsk_sdr::set_bb_gain(float value)
 {
     _bb_gain->set_k(value);
 }
